requirements.txt
Django==4.1.1
djangorestframework==3.14.0
mysqlclient==2.1.1
django-dotenv==1.4.2
drf-spectacular==0.24.1
python-dateutil==2.8.2